Moved shutdown information and variables from WSession to WBbsApp. Changed skey to take an int vs. unsigned char. More ongoing work on removing unneeded unsigned data type usages. Disabled remote.exe and remotes.dat support by default. Fixed display bug in voting booth.

+++ bbs/prot/zmutil.cpp
@@ -384,7 +384,7 @@ char * sname2(ZMState state)
 		"YTFile", "YTDataWait", "YTData", "YTEOF", "YTFin", "YRStart",
 		"YRDataWait", "YRData", "YREOF"
 	};
-	return names[(int)state];
+	return names[state];
 }
